feat: 초대 제외 보드관리 api 기능 완성

+++ services/boards.service.js
@@ -7,9 +7,9 @@ class BoardsService {
   boardsRepository = new BoardsRepository();
 
   // 보드 생성
-  createBoard = async (userId, name, color, description) => {
+  createBoard = async ({ userId, name, color, description }) => {
     try {
-      await this.boardsRepository.createOne({
+      await this.boardsRepository.createBoard({
         userId,
         name,
         color,
@@ -22,13 +22,13 @@ class BoardsService {
   };
 
   // 보드 수정
-  updateBoard = async (userId, boardId, name, color, description) => {
-    const targetBoard = await this.boardsRepository.findBoardById(boardId);
+  updateBoard = async ({ userId, boardId, name, color, description }) => {
+    const targetBoard = await this.boardsRepository.findBoardById({ boardId });
 
     if (!targetBoard) throw { code: 400, message: '보드가 존재하지 않습니다.' };
 
     try {
-      await this.boardsRepository.putOne({
+      await this.boardsRepository.updateBoard({
         userId,
         boardId,
         name,
@@ -43,27 +43,28 @@ class BoardsService {
   };
 
   // 보드 삭제
-  deleteBoard = async (boardId, userId) => {
-    await this.boardsRepository.findOneById(boardId);
+  deleteBoard = async ({ boardId, userId }) => {
+    const board = await this.boardsRepository.findBoardById({ boardId });
 
     if (!boardId) {
       throw { code: 400, message: '보드가 존재하지 않습니다.' };
     }
+    console.log(board);
 
-    if (boardId.userId !== userId) {
+    if (board.userId !== userId) {
       throw { code: 403, message: '권한이 없습니다.' };
     }
 
     try {
-      await this.boardsRepository.deleteBoard(boardId);
+      await this.boardsRepository.deleteBoard({ boardId, userId });
       return { code: 200, message: '보드를 삭제하였습니다.' };
     } catch (error) {
       throw { code: 400, message: '데이터 형식이 올바르지 않습니다.' };
     }
   };
 
-  inviteBoard = async (email, boardId) => {
-    const invitedUser = await this.boardsRepository.findByEmail(email);
+  inviteBoard = async ({ email, boardId }) => {
+    const invitedUser = await this.boardsRepository.findByEmail({ email });
 
     if (!invitedUser) {
       throw { code: 400, message: '초대할 유저가 존재하지 않습니다.' };
